1. Replaced the dialogs for loading and saving perspectives with file dialogs. 2. When a new perspective file is created, add a dynamic action for it.
Previously, I used an option dialog for selecting the perspective to load and
an input dialog for selecting the perspective to save. The reason for this was
that I intended to make the UI independent of the way the perspectives are
stored -- I planned to some day store them in the jEdit properties file rather
than XML files. But now that I'm quite satisfied with the XML files, so there's
no point in hiding this from the user, and file dialogs are more convenient
(and more standard) than the previous simple dialogs.