SF patch #1780421:  Mac thread safety on file open
Fixes a thread-safety issue that often hangs jEdit when opening files using
Mac machinery -- i.e. the Finder, drag-n-drop on the Dock icon, or the "open"
tool from the command line.

The problem was that Delegate.applicationOpenFiles(NSApplication, NSArray)
tried to open files from AWT-AppKit thread instead of AWT-EventQueue.  The only
way to recover was to Force Quit jEdit and restart it.