Properly implement a version control interface in ProjectViewer: . a new service (projectviewer.config.VersionControlService) is   available for plugins to implement. . the project properties now allows the user to choose a version   control service to be associated with the project. . only option panes related to the chosen VC service are shown to   the user. . IconComposer was refactored to be "per view", since different   projects may be using different VC services.
Currently, the internal IconComposer caching mechanism has been
removed; it doesn't seem to make CPU usage any different on my
machine, but I have a feeling it may increase memory usage; it
would have to be refactored anyway to support the new VC service
interface, so the old code won't be that much useful.

With the current changes, icon overlays are not available when the
root of the tree is not a project; to do that, the code would have
to look up the VC service for a given node every time the node is
repainted, by looking for the node's project and then for the project's
chosen service, which would probably not be very efficient. So I
chose to just not show overlays unless a project is the root.

The change also cleans up some of the API where it makes sense, by
making things that should not be public either private or
package-private, depending on the case.