Avoid hitting the VFS while building the search file path.
For some reason I don't exactly understand, the search subsystem
in jEdit fires a new I/O thread to calculate the search file set, and just
 waits for its completion from the AWT thread (effectively missing the 
point of using an I/O thread since the UI thread is still blocked).

Problem is that the VFS subsystem doesn't really provide a good way
for caching VFS sessions; so code that uses VFS needs to create and
destroy these sessions all the time. Since the create / destroy session
methods can use the UI thread (for password dialogs, for example), in
the above case this can lead to a deadlock.

So avoid hitting VFS in this case; the only thing the code doesn't do
anymore is check whether input files are readable, which shouldn't
really be an issue.