Fix the locking mess.
Having explicit locking was confusing and full of corner cases that
were not very easy to fix. Instead, move all locking to the internal
implementation of the tree nodes, so that code playing with the
project internals doesn't need to care about it.

This is probably a little less efficient than "lock - do lots of stuff
- unlock", since now the project lock will be locked / unlocked several
times when performing long operations, but it makes the code oh so much
easier to understand, and safer.