Allow VersionControlService to specify a file filter.
Instead of having VC-specific file filters always show up (for lack
of a better mechanism), allow the VersionControlService instance to
provide a file filter instance if it wants to. This filter will only
show up if the service is the one associated with the project on    
which action is being taken.