Switch to using 'Diff' and 'Patch' (from JDiffPlugin) for saving the list of changes in menus, and reapplying them on each start. This makes MenuEditor dependent on JDiffPlugin. Important: There are several small bugs in the implementation of 'Patch' in JDiffPlugin. Until they are fixed in SVN (and released), the file "jdiffplugin-patch.patch" can be used to fix them.
This change is done in order to avoid masking out changes to the menus that might be made in future versions of jEdit. Rather than keeping the resulting menus "fixed", the plugin just saves the "diff" and reapplies it each time. If jEdit introduces new menu items, the changes will be applied on top of them.