fix pb #3614106 Missing Hyperlinks to xref linkend= elements
Element declaration of elements in the root document
of a compound document were not handled correctly during parsing.

The stack of element declarations was no more in sync and attributes
with type ID in the root document were not recognised as such.
They were recognised only for xml:id or if there was a DTD.

Now the stack is correcly managed and "id" attributes
are thus recognised as such based on the XSD.